refactor(events): move CleanupBus from events to eventstest package (#95)
The production internal/domain/events package contained test_util.go which
imported "testing", a direct violation of ADR-021 ("test doubles live in
*test sub-packages"). This file leaked the testing package and its global
CLI flags into the production binary.

Relocate CleanupBus to internal/domain/events/eventstest/cleanup.go, which
is the correct home per ADR-021. Update 29 caller files (across both
internal/ and tests/integration/) to use eventstest.CleanupBus.

This is a precondition for the broader test-accessor refactor in #95: the
upcoming CI guard against "no testing import in production" cannot be added
until this violation is cleared.

No production behavior change. All tests pass.
